split thresholds loader
